change command api
return a list of notifications along with an updated world, where a
notification consists of a name and a message. this is _much_ more
flexible than `responses.'

also store client handles in nodes, not in a separate data structure.
this lets us clean up the server code a lot, at the cost of no longer
being able to use derived read/show instances for (de)serialization.